app-engine/requirements.txt
aiohttp==3.8.0
aiosignal==1.2.0
aniso8601==9.0.1
async-timeout==4.0.0
asynctest==0.13.0
attrs==21.2.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
decorator==5.1.0
Flask==2.0.2
flask-restplus==0.13.0
frozenlist==1.2.0
fsspec==2021.10.1
gcsfs==2021.10.1
google-api-core==2.2.0
google-auth==2.3.2
google-auth-oauthlib==0.4.6
google-cloud-appengine-logging==0.2.0
google-cloud-audit-log==0.2.0
google-cloud-core==2.1.0
google-cloud-dataproc==3.1.0
google-cloud-logging==2.6.0
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpcio==1.41.1
grpcio-status==1.41.1
idna==3.3
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
jsonschema==4.1.2
MarkupSafe==2.0.1
multidict==5.2.0
oauthlib==3.1.1
packaging==21.0
proto-plus==1.19.7
protobuf==3.19.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.11.0
pyparsing==3.0.3
pyrsistent==0.18.0
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
teradatasql==17.10.0.2
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
yarl==1.7.2
zipp==3.6.0
